reporters: rename xunit to junit, deprecate old xunit name
The xunit reporter has always emitted JUnit-format XML (testsuite/testcase
elements), not actual xUnit.net format. Fixes #4758.

- adds a new junit reporter with all the same logic
- turns xunit into a thin deprecation wrapper that extends JUnit
- the deprecation goes through process.emitWarning so its suppressable
  with --no-warnings and wont spam CI stderr like console.warn would
- updated docs and tests accordingly
