+++ docs/src/content/docs/reporters/junit.mdx
@@ -1,0 +1,16 @@
+---
+description: The JUnit test reporter for Mocha.
+title: JUnit
+---
+
+Alias: `JUnit`, `junit`
+
+The JUnit reporter outputs a JUnit-compatible XML document, often applicable in CI servers.
+
+By default, it will output to the console.
+To write directly to a file, use `--reporter-option output=filename.xml`.
+
+To specify custom report title, use `--reporter-option suiteName="Custom name"`.
+
+The reporter shows absolute file paths by default.
+To show relative paths instead, use `--reporter-option showRelativePaths`.
