+++ lib/mocha.js
@@ -298,8 +298,8 @@ Mocha.prototype.addFile = function (file) {
  * @throws {Error} if requested reporter cannot be loaded
  * @example
  *
- * // Use XUnit reporter and direct its output to file
- * mocha.reporter('xunit', { output: '/path/to/testspec.xunit.xml' });
+ * // Use JUnit reporter and direct its output to file
+ * mocha.reporter('junit', { output: '/path/to/testspec.junit.xml' });
  */
 Mocha.prototype.reporter = function (reporterName, reporterOptions) {
   if (typeof reporterName === "function") {
@@ -336,7 +336,7 @@ Mocha.prototype.reporter = function (reporterName, reporterOptions) {
     this._reporter = reporter;
   }
   this.options.reporterOption = reporterOptions;
-  // alias option name is used in built-in reporters xunit/tap/progress
+  // alias option name is used in built-in reporters junit/tap/progress
   this.options.reporterOptions = reporterOptions;
   return this;
 };
