+++ lib/reporters/index.js
@@ -12,6 +12,7 @@ exports.List = exports.list = require("./list");
 exports.Min = exports.min = require("./min");
 exports.Spec = exports.spec = require("./spec");
 exports.Nyan = exports.nyan = require("./nyan");
+exports.JUnit = exports.junit = require("./junit");
 exports.XUnit = exports.xunit = require("./xunit");
 exports.Markdown = exports.markdown = require("./markdown");
 exports.Progress = exports.progress = require("./progress");
